refactor: Flip note & sharing models dependencies
  We'll want to add dependencies to the `note` model package in the
  `sharing` model package so we can't have existing dependencies the
  other way around or we'll end up with circular dependencies.
